# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tv": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set watchlist or watched status for a TV series.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Upsert TV state",
                "parameters": [
                    {
                        "description": "TV state",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_tv.UpsertStateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tv/discover": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Browse the local TV cache with TMDB-style filters and sort. Each result is enriched with full TMDB detail via append_to_response.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Discover TV series",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "popularity.desc | first_air_date.desc | vote_average.desc | name.asc | max_watched_ep.desc | watchlisted_at.desc | watched_at.desc | ...",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated genre IDs (AND logic)",
                        "name": "with_genres",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated genre IDs to exclude",
                        "name": "without_genres",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by first air date year",
                        "name": "first_air_date_year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "First air date \u003e= (YYYY-MM-DD)",
                        "name": "first_air_date.gte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "First air date \u003c= (YYYY-MM-DD)",
                        "name": "first_air_date.lte",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Vote average \u003e=",
                        "name": "vote_average.gte",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Vote average \u003c=",
                        "name": "vote_average.lte",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Vote count \u003e=",
                        "name": "vote_count.gte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO 639-1 language code (e.g. en, th)",
                        "name": "with_original_language",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include adult content (default false)",
                        "name": "include_adult",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by softcore flag",
                        "name": "softcore",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated provider IDs",
                        "name": "with_watch_providers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO 3166-1 country code for watch provider filter",
                        "name": "watch_region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "watchlist | watching | waiting_next_ep | watched",
                        "name": "with_account_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated network IDs",
                        "name": "with_networks",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by anime flag",
                        "name": "is_anime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "TV series status (e.g. Returning Series, Ended)",
                        "name": "with_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "TV series type (e.g. Scripted, Documentary)",
                        "name": "with_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Next episode air date \u003e= (YYYY-MM-DD or RFC3339)",
                        "name": "with_next_episode_air_date.gte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Next episode air date \u003c= (YYYY-MM-DD or RFC3339)",
                        "name": "with_next_episode_air_date.lte",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_tv_TVResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tv/episodes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark specific episodes as watched for a TV series.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Add watched episodes",
                "parameters": [
                    {
                        "description": "Episodes to mark as watched",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_tv.UpsertEpisodesRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tv/random": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns TV series the user has no watchlist/watched record for, preferring series with an upcoming episode (next_episode_to_air, falling back to first_air_date) and falling back to popular titles. Results are randomized on every call.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Random TV suggestions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (echoed back, does not affect sampling; default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated account statuses to exclude, e.g. watchlist,watched,watching,waiting_next_ep",
                        "name": "without_status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results (default 20)",
                        "name": "total",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_tv_TVResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tv/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search TMDB for TV series. Results are enriched with genres and metadata from the local cache when available.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Search TV series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_tv_TVResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tv/states": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all TV series watchlist/history records for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "List TV states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_tv_TVStateResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tv/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns full TMDB detail for a TV series, with vote_average/vote_count and next_episode_to_air.air_date from the local cache.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Get TV detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TMDB TV id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_tv.TVResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_peera_movizius-go-service_internal_shared_response.Page-internal_tv_TVResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_tv.TVResponse"
                    }
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
        "github_com_peera_movizius-go-service_internal_shared_response.Page-internal_tv_TVStateResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_tv.TVStateResponse"
                    }
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
...
        "internal_tv.EpisodeInput": {
            "type": "object",
            "properties": {
                "episode_id": {
                    "type": "integer"
                },
                "episode_number": {
                    "type": "integer"
                },
                "season_number": {
                    "type": "integer"
                }
            }
        },
...
        "internal_tv.UpsertEpisodesRequest": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_tv.EpisodeInput"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "internal_tv.UpsertStateRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}